+++ chip-in-backend/src/main/java/dev/asif/chipinbackend/dto/ParticipantDTO.java
@@ -3,10 +3,12 @@ package dev.asif.chipinbackend.dto;
 import dev.asif.chipinbackend.model.ExpenseParticipant;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ParticipantDTO {
     private Long userId;
     private double percentage;
     private int shares;
-    private double shareAmount;
+    private BigDecimal shareAmount;
 }
